refactor and add workflow
