+++ .github/workflows/ucai-core-tests.yml
@@ -1,0 +1,59 @@
+name: tests
+
+on:
+  push:
+    branches:
+      - main
+    paths:
+      - unitycatalog-ai/**
+      - .github/workflows/ucai-*.yml
+  pull_request:
+    paths:
+      - unitycatalog-ai/**
+      - .github/workflows/ucai-*.yml
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.9'
+      - name: Install dependencies
+        run: |
+          pip install -r requirements/lint-requirements.txt
+      - name: Lint Python code with ruff
+        run: |
+          ruff check .
+          ruff format --check .
+      - name: Lint YAML files with yamllint
+        run: yamllint .
+
+  core_test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ['3.9', '3.10']
+        pydantic-version: [1, 2]
+    timeout-minutes: 20
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          pip install .[dev]
+          if [ ${{ matrix.pydantic-version }} == 1 ]; then
+            pip install 'pydantic<2'
+          else
+            pip install 'pydantic<3,>=2'
+          fi
+      - name: Run tests
+        run: |
+          pytest tests/
